s3_upload_api/requirements.txt
fastapi==0.115.0
pydantic==2.9.2
pydantic-settings==2.5.2
aioboto3==13.1.1
PyJWT==2.9.0
python-multipart==0.0.12

uvicorn~=0.35.0
# Dev / test
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
